refactor: move health.py to services/health.py
Consolidate the standalone health module into the services/ package.
Updates imports in server.py and tests/conftest.py.
